Fixed bug in CalculateVolumeWeightedStockPrice. Finished unit tests for CalculateVolumeWeightedStockPrice

+++ SuperSimpleStockMarket/Services/Interfaces/IStockService.cs
@@ -6,6 +6,6 @@ public interface IStockService
 {
     Decimal CalculateDividendYield(ref Stock stock, Decimal price);
     Decimal CalculatePERatio(ref Stock stock, Decimal price);
-    Decimal CalculateVolumeWeightedStockPrice(ref Stock stock, Int32 tradeIntervalMinutes = 5);
+    Decimal CalculateVolumeWeightedStockPrice(ref Stock stock, UInt16 tradeIntervalMinutes = 5);
     Boolean TryAddTrade(ref Stock stock, Trade trade);
 }
